Confrimed normalization of the calculation.
・シミュレーションでの計算の正常化を確認した。
・各モジュールのタイミングに関する問題が残ったため要修正。

+++ network_bench_verilog/_____00003aT0_00003a_00003aneural__network_00003aT0_00003a_00003alayer1_00003aT0_00003a_00003afunc0_00003aT00_00003a_00003amy__interpolator_00003aT10.v
@@ -6,6 +6,6 @@ module _____00003aT0_00003a_00003aneural__network_00003aT0_00003a_00003alayer1_0
    input signed[7:0] next__data;
    output signed[7:0] interpolated__value;
 
-   assign interpolated__value = (base + (($unsigned((next__data - base)) * remaining) >> 32'd4));
+   assign interpolated__value = (base + (($signed((next__data - base)) * remaining) >> 32'd4));
 
 endmodule
